+++ examples/apps/office_button_app.py
@@ -5,6 +5,7 @@ from typing import ClassVar
 from deepdiff import DeepDiff
 
 from hassette import AppConfig, AppSync, StateChangeEvent, entities, states
+from hassette import conditions as C
 from hassette import predicates as P
 from hassette.events import CallServiceEvent
 
@@ -28,7 +29,7 @@ class OfficeButtonApp(AppSync[OfficeButtonAppConfig]):
         self.bus.on_call_service(
             domain="light",
             handler=self.log_manual_light_service,
-            where=P.ServiceDataWhere.from_kwargs(entity_id=P.IsOrContains(self.app_config.office_light)),
+            where=P.ServiceDataWhere.from_kwargs(entity_id=C.IsOrContains(self.app_config.office_light)),
         )
 
         attributes = self.get_office_light().attributes