+++ src/hassette/__init__.py
@@ -8,9 +8,9 @@ from .core.resources.api.api import Api
 from .core.resources.app.app import App, AppSync, only_app
 from .core.resources.app.app_config import AppConfig, AppConfigT
 from .core.resources.base import Resource, Service
-from .core.resources.bus import predicates
 from .core.resources.bus.bus import Bus
 from .core.resources.bus.listeners import Listener, Subscription
+from .core.resources.bus.predicates import accessors, conditions, predicates
 from .core.resources.scheduler.classes import CronTrigger, IntervalTrigger, ScheduledJob
 from .core.resources.scheduler.scheduler import Scheduler
 from .core.resources.task_bucket import TaskBucket
@@ -18,7 +18,7 @@ from .enums import ResourceRole, ResourceStatus
 from .events import StateChangeEvent
 from .models import entities, states
 from .models.services import ServiceResponse
-from .types import AsyncHandler, Handler, Predicate, TriggerProtocol
+from .types import AsyncHandlerType, Predicate, TriggerProtocol
 
 logging.getLogger("hassette").addHandler(logging.NullHandler())
 
@@ -28,10 +28,9 @@ __all__ = [
     "AppConfig",
     "AppConfigT",
     "AppSync",
-    "AsyncHandler",
+    "AsyncHandlerType",
     "Bus",
     "CronTrigger",
-    "Handler",
     "Hassette",
     "HassetteConfig",
     "IntervalTrigger",
@@ -48,6 +47,8 @@ __all__ = [
     "Subscription",
     "TaskBucket",
     "TriggerProtocol",
+    "accessors",
+    "conditions",
     "context",
     "entities",
     "events",