+++ src/hassette/core/resources/bus/__init__.py
@@ -1,10 +1,4 @@
-from . import predicates
 from .bus import Bus
 from .listeners import Listener, Subscription
 
-__all__ = [
-    "Bus",
-    "Listener",
-    "Subscription",
-    "predicates",
-]
+__all__ = ["Bus", "Listener", "Subscription"]