+++ src/hassette/events/hass/hass.py
@@ -3,7 +3,7 @@ from dataclasses import dataclass, field
 from typing import Any, Generic, Self, TypeAlias
 
 from hassette import topics
-from hassette.const.misc import MISSING_VALUE
+from hassette.const import MISSING_VALUE
 from hassette.events.base import Event, HassPayload
 from hassette.models.states import StateT, try_convert_state