+++ src/hassette/types/types.py
@@ -1,4 +1,3 @@
-import typing
 from collections.abc import Awaitable, Callable, Mapping, Sequence
 from datetime import time
 from typing import Any, Protocol, TypeAlias, TypeVar
@@ -6,10 +5,7 @@ from typing import Any, Protocol, TypeAlias, TypeVar
 from typing_extensions import Sentinel, TypeAliasType
 from whenever import Date, PlainDateTime, Time, TimeDelta, ZonedDateTime
 
-if typing.TYPE_CHECKING:
-    from .events import Event
-
-EventT = TypeVar("EventT", bound="Event[Any]", contravariant=True)
+from .event import EventT
 
 V = TypeVar("V")  # value type from the accessor
 V_contra = TypeVar("V_contra", contravariant=True)
@@ -27,24 +23,10 @@ class Condition(Protocol[V_contra]):
     def __call__(self, value: V_contra, /) -> bool: ...
 
 
-class Handler(Protocol[EventT]):
-    """Protocol for defining event handlers."""
+class ComparisonCondition(Protocol[V_contra]):
+    """Protocol for a comparison condition callable that takes two values and returns a bool."""
 
-    def __call__(self, event: EventT) -> Awaitable[None] | None: ...
-
-
-class HandlerVariadic(Protocol[EventT]):
-    def __call__(self, event: EventT, *args: object, **kwargs: Any) -> Awaitable[None] | None: ...
-
-
-class AsyncHandler(Protocol[EventT]):
-    """Protocol for defining asynchronous event handlers."""
-
-    def __call__(self, event: EventT) -> Awaitable[None]: ...
-
-
-class AsyncHandlerVariadic(Protocol[EventT]):
-    def __call__(self, event: EventT, *args: object, **kwargs: Any) -> Awaitable[None]: ...
+    def __call__(self, old_value: V_contra, new_value: V_contra, /) -> bool: ...
 
 
 class TriggerProtocol(Protocol):
@@ -55,25 +37,15 @@ class TriggerProtocol(Protocol):
         ...
 
 
-AsyncHandlerType = TypeAliasType(
-    "AsyncHandlerType", AsyncHandler[EventT] | AsyncHandlerVariadic[EventT], type_params=(EventT,)
-)
-"""Alias for all valid async handler types."""
-
-HandlerType = TypeAliasType(
-    "HandlerType",
-    Handler[EventT] | HandlerVariadic[EventT] | AsyncHandler[EventT] | AsyncHandlerVariadic[EventT],
-    type_params=(EventT,),
-)
-"""Alias for all valid handler types."""
-
 KnownTypeScalar: TypeAlias = ZonedDateTime | PlainDateTime | Time | Date | None | float | int | bool | str
 """Alias for all known valid scalar state types."""
 
 KnownType: TypeAlias = KnownTypeScalar | Sequence[KnownTypeScalar] | Mapping[str, KnownTypeScalar]
 """Alias for all known valid state types."""
 
-ChangeType = TypeAliasType("ChangeType", None | Sentinel | V | Condition[V | Sentinel], type_params=(V,))
+ChangeType = TypeAliasType(
+    "ChangeType", None | Sentinel | V | Condition[V | Sentinel] | ComparisonCondition[V | Sentinel], type_params=(V,)
+)
 """Alias for types that can be used to specify changes in predicates."""
 
 JobCallable: TypeAlias = Callable[..., Awaitable[None]] | Callable[..., Any]